pyproject.toml
[project]
name = "medium-rag"
version = "0.1.0"
description = ""
authors = [
    {name = "Jill Villany",email = "jillian.villany@ibm.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "langchain (>=0.3.17,<0.4.0)",
    "langchain-openai (>=0.3.3,<0.4.0)",
    "langchain-pinecone (>=0.2.2,<0.3.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "black (>=25.1.0,<26.0.0)",
    "langchain-community (>=0.3.16,<0.4.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
